+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.5.0
+
+ - certificate "publisher" is now recognize automatic
+ - "publisher" configuration field is now deprecate
+ - added new flag/configuration-field "dontInstallCert"
+
 ## 2.4.2
 
  - Documentation update